Added ViewComponent Cocept to application
