+++ PernixMVC/ViewComponents/ShoppingCartViewComponent.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using PernixMVC.DataAccess.Repository.IRepository;
+using PernixMVC.Utility;
+using System.Security.Claims;
+
+namespace PernixMVC.ViewComponents
+{
+    public class ShoppingCartViewComponent : ViewComponent
+    {
+        private readonly IUnitOfWork _unitOfWork;
+        public ShoppingCartViewComponent(IUnitOfWork unitOfWork)
+        {
+            _unitOfWork = unitOfWork;
+        }
+
+        public async Task<IViewComponentResult> InvokeAsync()
+        {
+            var claimsIdentity = (ClaimsIdentity)User.Identity;
+            var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
+
+            if (claim != null)
+            {
+                if (HttpContext.Session.GetInt32(StaticDetails.SessionCart) == null)
+                {
+                    HttpContext.Session.SetInt32(StaticDetails.SessionCart,
+                    _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value).Count());
+                }
+
+                return View(HttpContext.Session.GetInt32(StaticDetails.SessionCart));
+            }
+            else
+            {
+                HttpContext.Session.Clear();
+                return View(0);
+            }
+        }
+
+    }
+
+}
