+++ PernixMVC/Views/Shared/Components/ShoppingCart/Default.cshtml
@@ -1,0 +1,4 @@
+﻿@model int
+
+
+<i class="bi bi-cart"></i> &nbsp; (@Model)
