+++ PernixMVC/Views/Shared/_Layout.cshtml
@@ -53,23 +53,13 @@
                                 </ul>
                             </li>
                         }
-                        @if(HttpContextAccessor.HttpContext.Session.GetInt32(StaticDetails.SessionCart) != null)
-                        {
+                       
                             <li class="nav-item">
                                 <a class="nav-link" asp-area="Customer" asp-controller="Cart" asp-action="Index">
-                                    <i class="bi bi-cart"></i> &nbsp;
-                                    (@HttpContextAccessor.HttpContext.Session.GetInt32(StaticDetails.SessionCart))
+                                   @await Component.InvokeAsync("ShoppingCart")
                                 </a>
                             </li>
-                        } else
-                        {
-                            <li class="nav-item">
-                                <a class="nav-link" asp-area="Customer" asp-controller="Cart" asp-action="Index">
-                                    <i class="bi bi-cart"></i> &nbsp;
-                                    (0)
-                                </a>
-                            </li>
-                        }
+                        
 
                     </ul>
                     <partial name="_LoginPartial" />
